refactor: restructure the masks package
Move core.py to masks package
Separate test package for the generic.py
